requirements.dev.txt
Django==4.2
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
drf-spectacular==0.26.5
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-allauth==0.50.0,
dj-rest-auth==2.2.4,
django-filter==23.3
redis==5.0.1
Pillow==10.1.0
phonenumbers==8.12.54
loguru==0.7.2
celery[redis]==5.2.7
six==1.16.0
django-tenants==3.6.1
squareup==37.1.0.20240604
drf-writable-nested==0.7.0
mailchimp-marketing==3.0.80
gunicorn==21.2.0
django-ses==4.2.0
flower==2.0.1
sentry-sdk[django]==2.20.0
deepdiff==8.2.0
channels-redis==4.2.1
daphne==4.1.2
Faker==37.0.0